Add sharemenu
