+++ lib/index.ts
@@ -139,6 +139,7 @@ export * from './src/patreonCharacters'
 export * from './src/random'
 export * from './src/randomFloat'
 export * from './src/rollFromTable'
+export * from './src/shareMenu'
 export * from './src/sortArray'
 export * from './src/terrain'
 export * from './src/tippy'
